Update Linux build
Fix some %ld message in log messages, and update the Linux sample
code to match recent changes in NufxLib and DiskImgLib.

Also, bump MDC version to 3.0.0 to match Windows version.

+++ linux/PackDDD.cpp
@@ -739,9 +739,9 @@ main(int argc, char** argv)
 
     printf("DDD Converter for Linux v1.0\n");
     printf("Copyright (C) 2003 by faddenSoft, LLC.  All rights reserved.\n");
-    long major, minor, bug;
+    int32_t major, minor, bug;
     Global::GetVersion(&major, &minor, &bug);
-    printf("Linked against DiskImg library v%ld.%ld.%ld\n",
+    printf("Linked against DiskImg library v%d.%d.%d\n",
         major, minor, bug);
 //    printf("Log file is '%s'\n", kLogFile);
     printf("\n");
